ml-service/requirements.txt
fastapi
uvicorn
pydantic
numpy
pandas
scikit-learn
joblib
httpx
python-dotenv
